+++ src/heavyedge/_wasserstein.pyx
@@ -8,9 +8,45 @@ import numpy as np
 
 cnp.import_array()
 
+
 @cython.boundscheck(False)
 @cython.wraparound(False)
-cpdef cnp.ndarray[cnp.float64_t, ndim=1] optimize_q(double[:] g):
+cpdef cnp.ndarray[cnp.float64_t, ndim=2] _quantile(double[:] x, double[:, :] Gs, cnp.int32_t[:] Ls, double[:] t):
+    cdef Py_ssize_t i, L, N = Gs.shape[0], M2 = t.shape[0]
+    cdef cnp.ndarray[cnp.float64_t, ndim=2] ret = np.empty((N, M2), dtype=np.float64)
+    for i in range(N):
+        L = Ls[i]
+        _quantile_interp(t, Gs[i, :L], x[:L], ret[i, :])
+    return ret
+
+
+cdef void _quantile_interp(double[:] t, double[:] G, double[:] x, double[:] out):
+    cdef Py_ssize_t i = 0, j = 0  # indices apply as: t[i], out[i], G[j], x[j]
+    cdef Py_ssize_t ii  # variable for emergency loop
+    cdef Py_ssize_t M2 = t.shape[0], L = G.shape[0]  # i: [0, M2), j: [0, L)
+    cdef double tval, G0, G1, x0, x1
+
+    out[i] = x[j]  # i = j = 0 here
+    for i in range(1, M2):
+        tval = t[i]
+        while j < L - 2 and G[j + 1] < tval:
+            j += 1
+
+            G0 = G[j]
+            G1 = G[j + 1]
+            x0 = x[j]
+            x1 = x[j + 1]
+
+        if tval < G[j + 1]:
+            out[i] = x0 + (x1 - x0) / (G1 - G0) * (tval - G0)
+        else:
+            # G[-1] < 0 for numerical reason.
+            out[i] = x1
+
+
+@cython.boundscheck(False)
+@cython.wraparound(False)
+cpdef cnp.ndarray[cnp.float64_t, ndim=1] _optimize_q(double[:] g):
     # M = number of probability-space grid, N = number of distance-space grid
     cdef Py_ssize_t i, j, idx
     cdef double[:] y_vals = np.unique(g)