+++ CHANGELOG.md
@@ -18,6 +18,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Deprecated
 
 - `RawProfileBase.all_profiles()` method is deprecated. Directly iterate `profiles()` method instead.
+- `wdist()` function is deprecated. Use HeavyEdge-Distance package instead.
 
 ## [1.5.0] - 2025-10-21